requirements.txt
flask==3.0.3
twilio==9.2.3
google-auth==2.29.0
google-api-python-client==2.128.0
gunicorn==22.0.0
